New test case for librom. Removed test from actual module
